+++ .test.js
@@ -1,4 +1,6 @@
-const baseConversion = require('./baseConversion.js');
+import { jest } from '@jest/globals';
+import { convertNumberToBase } from "./utils";
+
 
 const BASE_CONVERSION_TESTS = [{
         str: '64',
@@ -23,8 +25,13 @@ const BASE_CONVERSION_TESTS = [{
     }
 ]
 
+const baseConversion = jest.fn( (str, startBase, changeBase, targetLen = 0) => {
+    const res = convertNumberToBase(str, startBase, changeBase);
+    return convertNumberToBase(res, changeBase, startBase, targetLen);
+});
+
 BASE_CONVERSION_TESTS.forEach(e => {
     test(`Converts ${e.str} from Base-${e.startBase} to Base-${e.changeBase} then back to Base-${e.startBase}, becoming ${e.str} again.`, () => {
-        expect(baseConversion.baseConversion(e.str, e.startBase, e.changeBase, e.padToLen)).toBe(e.str);
+        expect(baseConversion(e.str, e.startBase, e.changeBase, e.padToLen)).toBe(e.str);
     });
 });
